script/requirements.txt
tabulate
scikit-surprise